+++ prototyping_inference_engine/api/atom/term/__init__.py
@@ -1,0 +1,3 @@
+from prototyping_inference_engine.api.atom.term.function_term import FunctionTerm
+
+__all__ = ["FunctionTerm"]